orderbooks: convert base classes to TS
