--- src/orderbooks/L2Point.js
@@ -1,9 +1,0 @@
-class L2Point {
-  constructor(price, size, timestamp) {
-    this.price = price;
-    this.size = size;
-    this.timestamp = timestamp;
-  }
-}
-
-module.exports.L2Point = L2Point;
